requirements.txt
bleach==5.0.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
docutils==0.18.1
future==0.18.2
gitdb2==2.0.6
githubrelease==1.5.8
GitPython==2.1.11
idna==3.3
importlib-metadata==4.11.4
latexcodec==2.0.1
LinkHeader==0.4.3
lxml==4.9.0
pkg-resources==0.0.0
pkginfo==1.8.2
publish-doconce==1.1.4
pybtex==0.24.0
Pygments==2.12.0
python-Levenshtein-wheels==0.13.2
PyYAML==6.0
readme-renderer==35.0
requests==2.27.1
requests-toolbelt==0.9.1
six==1.12.0
smmap==5.0.0
smmap2==3.0.1
tqdm==4.64.0
twine==1.12.1
typing-extensions==4.2.0
urllib3==1.26.9
webencodings==0.5.1
zipp==3.8.0
